requirements.txt
aiogram
python-dotenv
Telethon>=1.40.0
asyncpg
psycopg2-binary>=2.9.9 

# Додаткові бібліотеки, які можуть знадобитися пізніше:
# asyncpg
# aiofiles
# requests
# flask
